test: Restructure test namespaces
