+++ tests/Bot.Integration.Tests/Clockify/ClockifyReadApiTest.cs
@@ -3,7 +3,6 @@ using System;
 using System.Collections.Generic;
 using System.Threading.Tasks;
 using Bot.Clockify.Client;
-using Bot.Services;
 using FluentAssertions;
 using Microsoft.Bot.Schema;
 using Xunit;
